some refactoring & code cleaning
so i found out that Gnometris (now Quadrapassel) has a "Bastard" mode, which
is the "unique" feature i was going to add to tetrellis to make it different
from other tetris clones.  i guess not.  a lot of other tetris clones
probably have this feature already as well.

so, new goal: tighten & refactor this thing through & through and make it
the cleanest nicest C code i've ever built.  it'll be a good exercise for
me, on top of the SDL practise i've been getting with it.